refactor: 패키지명 복수형에서 단수형으로 수정
