+++ app/src/main/AndroidManifest.xml
@@ -15,10 +15,10 @@
         android:theme="@style/Theme.ParsingGithubRepositories"
         tools:targetApi="31">
         <activity
-            android:name=".ui.views.DetailActivity"
+            android:name=".ui.view.DetailActivity"
             android:exported="false" />
         <activity
-            android:name=".ui.views.HomeActivity"
+            android:name=".ui.view.HomeActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
