+++ app/src/main/res/layout/activity_home.xml
@@ -9,13 +9,13 @@
 
         <variable
             name="viewModel"
-            type="com.hongstudio.parsing_github_repositories.ui.viewmodels.HomeViewModel" />
+            type="com.hongstudio.parsing_github_repositories.ui.viewmodel.HomeViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".ui.views.HomeActivity">
+        tools:context=".ui.view.HomeActivity">
 
         <LinearLayout
             android:id="@+id/searchLayout"
